refactor common entities props and methods
